+++ staff-ui/src/pages/blog/BlogPostEdit.tsx
@@ -1,0 +1,11 @@
+import { Edit } from 'react-admin'
+
+import { BlogPostForm } from './BlogPostForm'
+
+export default function BlogPostEdit() {
+  return (
+    <Edit>
+      <BlogPostForm />
+    </Edit>
+  )
+}