+++ staff-ui/src/pages/portfolio/PortfolioCreate.tsx
@@ -1,0 +1,11 @@
+import { Create } from 'react-admin'
+
+import { PortfolioForm } from './PortfolioForm'
+
+export default function PortfolioCreate() {
+  return (
+    <Create>
+      <PortfolioForm />
+    </Create>
+  )
+}